# swagger-example/swagger_config.yml
openapi: 3.0.3
info:
  title: Task Organizer - with OpenAPI 3.0
  description: |-
    Organize and maintain tasks
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:5000
tags:
  - name: swagger example
    description: Organize and maintain tasks
paths:
  /tasks:
    get:
      tags:
          - tasks
      summary: Returns a list of tasks.
      description: Optional extended description in Markdown.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
        '405':
          description: Invalid Input
  /task:
    post:
      tags:
        - task
      summary: Adding task to list
      consumes:
        - application/x-www-form-urlencoded
      requestBody:
        description: Created task object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
          application/xml:
            schema:
              $ref: '#/components/schemas/Task'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: OK
  /task/{taskname}:
    get:
      tags:
        - task name
      summary: Returns a task by name.
      parameters:
        - name: taskname
          in: path
          required: true
          description: Name of the task to be obtained
          schema:
            type: string
      responses:
        '200':
          description: Successfully fetched task
    delete:
      tags:
        - task name
      summary: Delete a task by name.
      parameters:
        - name: taskname
          in: path
          description: Name of the task to be deleted
          required: true
          schema:
            type: string
      responses:
        '204':
          description: no content
        '400':
          description: Invalid username supplied
        '404':
          description: User not found

components:
  schemas:
    Task:
      type: object
      properties:
        name:
          type: string
          example: Task name
        description:
          type: string
          description: Description of the task